feat: refactored loader slightly
