+++ votes/loader/src/logger.ts
@@ -1,0 +1,18 @@
+import winston from 'winston';
+
+const logger = winston.createLogger({level: 'info',
+  format: winston.format.json(),
+  defaultMeta: { service: 'file-processor'},
+  transports: [
+    new winston.transports.File({filename: 'error.log', level: 'error'}),
+    new winston.transports.File({filename: 'combined.log'})
+  ],
+})
+
+if (process.env.NODE_ENV !== 'prod') {
+  logger.add(new winston.transports.Console({
+    format: winston.format.simple()
+  }))
+}
+
+export default logger;
